+++ src/algorithms/expval.jl
@@ -1,22 +1,26 @@
 """
     expectation_value(ψ, O, [environments])
     expectation_value(ψ, inds => O)
+    expectation_value(ψ, (mpo, site => O), [environments])
 
 Compute the expectation value of an operator `O` on a state `ψ`. 
 Optionally, it is possible to make the computations more efficient by also passing in
 previously calculated `environments`.
 
 In general, the operator `O` may consist of an arbitrary MPO `O <: AbstractMPO` that acts
-on all sites, or a local operator `O = inds => operator` acting on a subset of sites. 
-In the latter case, `inds` is a tuple of indices that specify the sites on which the operator
-acts, while the operator is either a `AbstractTensorMap` or a `FiniteMPO`.
+on all sites, a local operator `O = inds => operator` acting on a subset of sites, or a local MPO tensor
+acting on a site within a network whose environment is determined by another MPO `mpo`.
+In the second case, `inds` is a tuple of indices that specify the sites on which the operator
+acts, while the operator is either a `AbstractTensorMap` or a `FiniteMPO`. In the latter case,
+the operator is a `AbstractTensorMap` that acts on the physical space of a single site.
 
 # Arguments
 * `ψ::AbstractMPS` : the state on which to compute the expectation value
-* `O::Union{AbstractMPO,Pair}` : the operator to compute the expectation value of. 
-    This can either be an `AbstractMPO`, or a pair of indices and local operator..
+* `O::Union{AbstractMPO,Pair,AbstractTensorMap}` : the operator to compute the expectation value of. 
+    This can either be an `AbstractMPO`, a pair of indices and local operator, or a local MPO tensor
+    represented as a `AbstractTensorMap`.
 * `environments::AbstractMPSEnvironments` : the environments to use for the calculation. If not given, they will be calculated.
-
+    Depending on the type of `O`, these will be the environments of the operator `O` or the MPO `mpo`.
 # Examples
 ```jldoctest
 julia> ψ = FiniteMPS(ones(Float64, (ℂ^2)^4));
@@ -128,11 +132,22 @@ function expectation_value(
         ψ::InfiniteMPS, H::InfiniteMPOHamiltonian,
         envs::AbstractMPSEnvironments = environments(ψ, H)
     )
-    return sum(1:length(ψ)) do i
+    return sum(1:length(ψ)) do site
         return contract_mpo_expval(
-            ψ.AC[i], envs.GLs[i], H[i][:, 1, 1, end], envs.GRs[i][end]
+            ψ.AC[site], envs.GLs[site], H[site][:, 1, 1, end], envs.GRs[site][end]
         )
     end
+end
+
+# MPO tensor
+#-----------
+function expectation_value(
+        ψ::InfiniteMPS, mpo_pair::Tuple{InfiniteMPO, Pair{Int, <:MPOTensor}},
+        envs::AbstractMPSEnvironments = environments(ψ, first(mpo_pair))
+    )
+    O_mpo, (site, O) = mpo_pair
+    return contract_mpo_expval(ψ.AC[site], envs.GLs[site], O, envs.GRs[site]) /
+        expectation_value(ψ, O_mpo, envs)
 end
 
 # DenseMPO
@@ -194,8 +209,7 @@ function expectation_value(
     )
     ens = zeros(scalartype(ψ), length(ψ))
     for i in 1:length(ψ)
-        operator = AC_hamiltonian(i, ψ, O, ψ, envs)
-        ens[i] = dot(ψ.AC[i], operator * ψ.AC[i])
+        ens[i] = dot(ψ.AC[i], AC_projection(i, ψ, O, ψ, envs))
     end
     n = norm(ψ.AC[end])^2
     return sum(ens) / (n * length(ψ))