+++ src/algorithms/groundstate/vumps.jl
@@ -57,6 +57,7 @@ function dominant_eigsolve(
     )
     log = IterLog("VUMPS")
     iter = 0
+    mps = copy(mps)
     ϵ = calc_galerkin(mps, operator, mps, envs)
     alg_environments = updatetol(alg.alg_environments, iter, ϵ)
     recalculate!(envs, mps, operator, mps; alg_environments.tol)
@@ -110,7 +111,7 @@ function localupdate_step!(
     mps = state.mps
     src_Cs = mps isa Multiline ? eachcol(mps.C) : mps.C
     src_ACs = mps isa Multiline ? eachcol(mps.AC) : mps.AC
-    ACs = similar(mps.AC)
+    ACs = mps.AL
     dst_ACs = mps isa Multiline ? eachcol(ACs) : ACs
 
     tforeach(eachsite(mps), src_ACs, src_Cs; scheduler) do site, AC₀, C₀
@@ -153,7 +154,12 @@ end
 
 function gauge_step!(it::IterativeSolver{<:VUMPS}, state, ACs::AbstractVector)
     alg_gauge = updatetol(it.alg_gauge, state.iter, state.ϵ)
-    return InfiniteMPS(ACs, state.mps.C[end]; alg_gauge.tol, alg_gauge.maxiter)
+    mps = gaugefix!(
+        state.mps, ACs, state.mps.C[end];
+        alg_gauge.tol, alg_gauge.maxiter, order = :R
+    )
+    mul!.(mps.AC, mps.AL, mps.C)
+    return mps
 end
 function gauge_step!(it::IterativeSolver{<:VUMPS}, state, ACs::AbstractMatrix)
     alg_gauge = updatetol(it.alg_gauge, state.iter, state.ϵ)