+++ src/operators/mpohamiltonian.jl
@@ -32,6 +32,7 @@ struct MPOHamiltonian{TO <: JordanMPOTensor, V <: AbstractVector{TO}} <: Abstrac
     W::V
 end
 OperatorStyle(::Type{<:MPOHamiltonian}) = HamiltonianStyle()
+TensorKit.storagetype(::Type{MPOHamiltonian{O, V}}) where {O, V} = storagetype(O)
 
 const FiniteMPOHamiltonian{O <: MPOTensor} = MPOHamiltonian{O, Vector{O}}
 Base.isfinite(::Type{<:FiniteMPOHamiltonian}) = true
@@ -437,11 +438,11 @@ function FiniteMPOHamiltonian(lattice::AbstractArray{<:VectorSpace}, local_opera
 
     # construct the sparse MPO
     T = _find_tensortype(nonzero_opps)
-    E = scalartype(T)
     S = spacetype(T)
 
     # avoid using one(S)
     P = first(lattice)
+    P = P isa ProductSpace ? P[length(P)] : P
     _rightunit = rightunitspace(P)
     @assert _rightunit == leftunitspace(P) "only diagonal hamiltonians allowed"
 
@@ -465,7 +466,7 @@ function FiniteMPOHamiltonian(lattice::AbstractArray{<:VectorSpace}, local_opera
     end
 
     # construct the tensor
-    TW = jordanmpotensortype(S, E)
+    TW = jordanmpotensortype(T)
     Os = map(1:length(lattice)) do site
         V = virtualsumspaces[site] * lattice[site] ←
             lattice[site] * virtualsumspaces[site + 1]
@@ -476,7 +477,7 @@ function FiniteMPOHamiltonian(lattice::AbstractArray{<:VectorSpace}, local_opera
             key_R = key_R′ == 0 ? length(virtualsumspaces[site + 1]) : key_R′
             O[key_L, 1, 1, key_R] += if o isa Number
                 iszero(o) && continue
-                τ = BraidingTensor{E}(eachspace(O)[key_L, 1, 1, key_R])
+                τ = BraidingTensor{scalartype(TW)}(eachspace(O)[key_L, 1, 1, key_R])
                 isone(o) ? τ : τ * o
             else
                 o
@@ -520,7 +521,6 @@ function InfiniteMPOHamiltonian(lattice′::AbstractArray{<:VectorSpace}, local_
 
     # construct the sparse MPO
     T = _find_tensortype(nonzero_opps)
-    E = scalartype(T)
     S = spacetype(T)
 
     # construct the virtual spaces
@@ -531,6 +531,7 @@ function InfiniteMPOHamiltonian(lattice′::AbstractArray{<:VectorSpace}, local_
     )
     # avoid using one(S)
     P = first(lattice)
+    P = P isa ProductSpace ? P[length(P)] : P
     _rightunit = rightunitspace(P)
     @assert _rightunit == leftunitspace(P) "only diagonal hamiltonians allowed"
 
@@ -588,7 +589,7 @@ function InfiniteMPOHamiltonian(lattice′::AbstractArray{<:VectorSpace}, local_
     end
 
     # construct the tensor
-    TW = jordanmpotensortype(S, E)
+    TW = jordanmpotensortype(T)
     Os = map(1:length(lattice)) do site
         V = virtualsumspaces[site - 1] * lattice[site] ←
             lattice[site] * virtualsumspaces[site]
@@ -599,7 +600,7 @@ function InfiniteMPOHamiltonian(lattice′::AbstractArray{<:VectorSpace}, local_
             key_R = key_R′ == 0 ? length(virtualspaces[site]) : key_R′
             O[key_L, 1, 1, key_R] += if o isa Number
                 iszero(o) && continue
-                τ = BraidingTensor{E}(eachspace(O)[key_L, 1, 1, key_R])
+                τ = BraidingTensor{scalartype(TW)}(eachspace(O)[key_L, 1, 1, key_R])
                 isone(o) ? τ : τ * o
             else
                 o
@@ -824,8 +825,7 @@ function Base.:*(H::FiniteMPOHamiltonian, mps::FiniteMPS)
         )
     )
     # left to middle
-    U = ones(scalartype(H), left_virtualspace(H, 1))
-    @plansor a[-1 -2; -3 -4] := A[1][-1 2; -3] * H[1][1 -2; 2 -4] * conj(U[1])
+    @plansor a[-1 -2; -3 -4] := A[1][-1 1; -3] * removeunit(H[1], 1)[-2; 1 -4]
     Q, R = qr_compact!(a)
     A′[1] = TensorMap(Q)
 
@@ -836,8 +836,7 @@ function Base.:*(H::FiniteMPOHamiltonian, mps::FiniteMPS)
     end
 
     # right to middle
-    U = ones(scalartype(H), right_virtualspace(H, N))
-    @plansor a[-1 -2; -3 -4] := A[end][-1 2; -3] * H[end][-2 -4; 2 1] * U[1]
+    @plansor a[-1 -2; -3 -4] := A[end][-1 1; -3] * removeunit(H[end], 4)[-2 -4; 1]
     L, Q = lq_compact!(a)
     A′[end] = transpose(TensorMap(Q), ((1, 3), (2,)))