+++ test/runtests.jl
@@ -11,5 +11,15 @@ init_code = quote
     using .TestSetup
 end
 
+# only run CUDA if on buildkite
+is_buildkite = get(ENV, "BUILDKITE", "false") == "true"
+is_buildkite && filter!(startswith("cuda") ∘ first, testsuite)
+
+# only run CUDA/cuTENSOR if available
+using CUDA, cuTENSOR
+(CUDA.functional() && cuTENSOR.has_cutensor()) ||
+    filter!(!(startswith("cuda") ∘ first), testsuite)
+
+# run tests
 args = parse_args(ARGS)
 runtests(MPSKit, args; testsuite, init_code)