+++ docs/src/changelog.md
@@ -17,7 +17,7 @@ When making changes to this project, please update the "Unreleased" section with
 
 When releasing a new version, move the "Unreleased" changes to a new version section with the release date.
 
-## [Unreleased](https://github.com/QuantumKitHub/MPSKit.jl/compare/v0.13.9...HEAD)
+## [Unreleased](https://github.com/QuantumKitHub/MPSKit.jl/compare/v0.13.10...HEAD)
 
 ### Added
 
@@ -28,6 +28,26 @@ When releasing a new version, move the "Unreleased" changes to a new version sec
 ### Removed
 
 ### Fixed
+
+## [0.13.10](https://github.com/QuantumKitHub/MPSKit.jl/compare/v0.13.9...v0.13.10) - 2026-02-26
+
+### Added
+
+- `expectation_value` for local MPO tensors
+  ([#327](https://github.com/QuantumKitHub/MPSKit.jl/pull/327))
+- `Base.copy` for MPS types now performs a deep copy
+  ([#387](https://github.com/QuantumKitHub/MPSKit.jl/pull/387))
+
+### Changed
+
+- `entropy` can now also be called directly on a spectrum (singular value vector)
+  ([#377](https://github.com/QuantumKitHub/MPSKit.jl/pull/377))
+- Updated compat bounds to remove broken package versions
+
+### Fixed
+
+- Fixed `Adapt` extension for GPU support
+  ([#389](https://github.com/QuantumKitHub/MPSKit.jl/pull/389))
 
 ## [0.13.9](https://github.com/QuantumKitHub/MPSKit.jl/compare/v0.13.8...v0.13.9) - 2026-02-03